msm8916-common: Provide android.hardware.radio@1.1 on vendor
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.radio@1.1::IRadio/slot1": error code: 0x20
I hwservicemanager: Tried to start android.hardware.radio@1.1::IRadio/slot1 as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.